+++ godot/src/Consts.cs
@@ -2,6 +2,6 @@ namespace DefectOverhaul;
 
 public static class Consts {
     public const string ModId = "DefectOverhaul";
-    public const string ModVersion = "0.100.3";
-    public const string Sts2Version = "0.107.0";
+    public const string ModVersion = "0.101.0";
+    public const string Sts2Version = "0.108.0";
 }